Working on creatig the player struct from json Need to work on invenory and also add a starting inventory attribute for each class and then continue starting new game intro also the reason for setting level 0 is becuase \ we complete the gameintro we set to 1 and it can be a signifyer of newgame intro complete

+++ C_Programs/rpg/Headers/rpgStructs.h
@@ -20,17 +20,17 @@ typedef struct{
 
     item *inventory;
 
-    float maxHealth;
-    float health;
+    double maxHealth;
+    double health;
 
-    float maxStrength;
-    float strength;
+    double maxStrength;
+    double strength;
 
-    int speed;
-    int maxSpeed;
+    double speed;
+    double maxSpeed;
 
-    float mana;
-    float maxMana;
+    double mana;
+    double maxMana;
 
 
 }Player;
